add options "target" and "method"
use `target` to provide an object you want the transformation to be merged into. merge happens in order: target, picked, mapped. after that, the result will be handled by the `parse` transformer.

use `method` to define which lodash function you want to use to "join" target, picked and mapped. default is "assign"